+++ crates/rils_frontend/src/standard_library.rs
@@ -233,6 +233,7 @@ pub(crate) fn resolve_type_pattern(pattern: rils_builtins::TypePattern) -> Type 
         TypePattern::String => Type::String,
         TypePattern::F32 => Type::Float(crate::types::FloatType::F32),
         TypePattern::F64 => Type::Float(crate::types::FloatType::F64),
+        TypePattern::U32 => Type::Integer(crate::types::IntegerType::U32),
         TypePattern::Usize => Type::USIZE,
         TypePattern::Named { path, arguments } => Type::Named {
             name: path.into(),